implemented changes for obtaining metadata on maps with data not available through discovery API. setup socrata app token
there are datasets which appear only as maps, and not datasets on the NB socrata when queried using the discovery API. however, data is publicly queryable through the API endpoint found on the map entry on socrata.  the api endpoint for the map is NOT the same as the api endpoint for the data it is based on. the link is also hidden behind a javascript toggle, and therefore can not be extracted using the htmlextractor transformer.

to resolve this, I have implemented a python caller which runs a headless chrome browser to "click" the necessary links to access the true api endpoint. it then uses the beautifulsoup library to parse the html and obtain the API endpoint.

you will need to install selenium to the default fme python environment for this to work. see the following links

https://docs.safe.com/fme/html/FME_Desktop_Documentation/FME_Workbench/Workbench/Installing-Python-Packages.htm

https://docs.safe.com/fme/html/FME_Server_Documentation/AdminGuide/Using-Python-with-FME_Server.htm

the browser can be changed to firefox, internet explorer, etc and must be updated with the web browser - in the future will probably include a portable chrome browser in the repo so we do not have to update it.

Also, obtained a temporary app token so our requests are not throttled. we will need to get an organization app token.